requirements.txt
django 
djangorestframework
django-extensions
requests